fix types

+++ internal/domain/warehouse.go
@@ -1,10 +1,10 @@
 package domain
 
 type Warehouse struct {
-	ID                 int    `json:"id"`
-	Address            string `json:"address"`
-	Telephone          string `json:"telephone"`
-	WarehouseCode      string `json:"warehouse_code"`
-	MinimumCapacity    int    `json:"minimum_capacity"`
-	MinimumTemperature int    `json:"minimum_temperature"`
+	ID                 int      `json:"id"`
+	Address            string   `json:"address"`
+	Telephone          string   `json:"telephone"`
+	WarehouseCode      string   `json:"warehouse_code"`
+	MinimumCapacity    int      `json:"minimum_capacity"`
+	MinimumTemperature float64  `json:"minimum_temperature"`
 }
